Rename all usages of openfoam and it's submodules to match new names
