+++ ApertureMapModelTools/__init__.py
@@ -17,7 +17,7 @@ from .__core__ import files_from_directory, load_infile_list
 from .__core__ import calc_percentile, calc_percentile_num, get_data_vect
 from . import data_processing
 from . import RunModel
-from . import OpenFoam
+from . import openfoam
 
 
 __version__ = '0.0.2'
