+++ test/unit/TestOpenFoam/TestParallelMeshGen.py
@@ -11,11 +11,11 @@ import os
 import pytest
 import sys
 import scipy as sp
-import ApertureMapModelTools.OpenFoam.__ParallelMeshGen__ as pmg_submodule
-from ApertureMapModelTools.OpenFoam.__ParallelMeshGen__ import DataFieldRegion
-from ApertureMapModelTools.OpenFoam.__ParallelMeshGen__ import BlockMeshRegion
-from ApertureMapModelTools.OpenFoam.__ParallelMeshGen__ import MergeGroup
-from ApertureMapModelTools.OpenFoam import ParallelMeshGen
+import ApertureMapModelTools.openfoam.parallel_mesh_gen as pmg_submodule
+from ApertureMapModelTools.openfoam.parallel_mesh_gen import DataFieldRegion
+from ApertureMapModelTools.openfoam.parallel_mesh_gen import BlockMeshRegion
+from ApertureMapModelTools.openfoam.parallel_mesh_gen import MergeGroup
+from ApertureMapModelTools.openfoam import ParallelMeshGen
 
 
 @pytest.mark.xfail(sys.platform == 'win32', reason="OpenFoam doesn't natively run on Windows")
